requirements.txt
hydra-core~=1.3.2
omegaconf~=2.3.0
transformers~=4.52.3
datasets~=3.6.0
peft~=0.15.2
torch~=2.7.0
evaluate~=0.4.3
pandas~=2.2.3
